requirements.txt
langchain
langchain-classic
langchain-community
langchain-chroma
langchain-ollama
langchain-google-genai
chromadb
python-dotenv
rank-bm25
streamlit
